sandbox/pyproject.toml
[project]
name = "sphinx_hosting_demo"
version = "0.1.0"
authors = [
    {name = "Caltech IMSS ADS", email = "imss-ads-staff@caltech.edu"},
]
maintainers = [
    {name = "Chris Malek", email = "cmalek@caltech.edu"},
]
requires-python = ">= 3.11"
description = "Demo app for django-sphinx-hosting"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 4.1",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.0",
    "Topic :: Documentation :: Sphinx",
    "Topic :: Software Development :: Documentation",
]
dependencies = [
    "django-sphinx-hosting",
    "django-environ==0.11.2",
    "django-xff==1.3.0",
    "django-compressor>=4.5.1",
    "django-crequest==2018.5.11",
    "django-debug-toolbar>=4.4.2",
    "gunicorn >= 22.0.0",
    "elasticsearch==7.13.3",
    "ipython>=7.27.0",
    "mysqlclient==2.2.6",
    "redis==5.2.1",
    "structlog==24.2.0",
    "tabulate==0.9.0",
    "django-sass-processor==1.4.1",
    "libsass==0.23.0",
]

[tool.uv.sources]
django-sphinx-hosting = {path = "./django-sphinx-hosting.tar.gz"}

[tool.uv]
python-preference = "only-system"

[tool.setuptools]
# ...
# By default, include-package-data is true in pyproject.toml, so you do
# NOT have to specify this line.
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]